Fixed download, split file Download.cs, also added check approciate network and changed from Contact.ID check to Contact.MachineId check in BucketList.AddContact

+++ DiscoveryService/Messages/KademliaRequests/DownloadFileRequest.cs
@@ -73,7 +73,7 @@ namespace LUC.DiscoveryService.Messages.KademliaRequests
 
             if(downloadResponse?.Chunk?.Length > 0)
             {
-                CountDownloadedBytes += (UInt32)downloadResponse.Chunk.Length;
+                CountDownloadedBytes += (UInt64)downloadResponse.Chunk.Length;
 
                 //remove next downloaded chunk
                 Range.NumsUndownloadedChunk.RemoveAt(index: 0);
